// src/entity/discipline/models/field-label.ts
export enum FieldLabel {
  Name = 'Название дисциплины',
  Description = 'Описание дисциплины',
  presentationLink = 'Ссылка на презентацию',
  modeusLink = 'Ссылка на расписание в Modeus',
  Module = 'Название модуля',
  Format = 'Формат',
}